+++ apps/scan/src/main/scala/com/daml/network/scan/admin/api/client/commands/HttpScanAppClient.scala
@@ -763,7 +763,9 @@ object HttpScanAppClient {
     }
   }
   object GetAggregatedRounds
-      extends InternalBaseCommand[http.GetAggregatedRoundsResponse, ScanAggregator.RoundRange] {
+      extends InternalBaseCommand[http.GetAggregatedRoundsResponse, Option[
+        ScanAggregator.RoundRange
+      ]] {
     override def submitRequest(
         client: http.ScanClient,
         headers: List[HttpHeader],
@@ -776,9 +778,9 @@ object HttpScanAppClient {
 
     override def handleOk()(implicit decoder: TemplateJsonDecoder) = {
       case http.GetAggregatedRoundsResponse.OK(response) =>
-        Right(ScanAggregator.RoundRange(response.start, response.end))
-      case http.GetAggregatedRoundsResponse.NotFound(err) =>
-        Left(err.error)
+        Right(Some(ScanAggregator.RoundRange(response.start, response.end)))
+      case http.GetAggregatedRoundsResponse.NotFound(_) =>
+        Right(None)
     }
   }
   case class ListRoundTotals(start: Long, end: Long)
@@ -801,6 +803,7 @@ object HttpScanAppClient {
         Right(response.entries)
     }
   }
+
   case class ListRoundPartyTotals(start: Long, end: Long)
       extends InternalBaseCommand[
         http.ListRoundPartyTotalsResponse,